refactor(types.ts): allow handler functions to return any
Currently handler functions that are passed to `Webhooks.on` can only
return `void` or `Promise<void>`. The return values from these functions
are only used to wait for all handlers for a given webhook event to
finish before the promise returned from `Webhooks.receive` resolves (or
rejects if any of the handler functions error out).

This updates the type signature to allow handler functions to return
anything.

Since this type signature was used in both `Webhooks.on` and
`Webhooks.removeListener`, this PR adds named type called
`HandlerFunction<E, T>` in `types.ts` and adds test cases to
`typescript-validate.ts` for the more lax types.