refactor: use test promise wait over done callback
This changes the `done` callback functionality that tells a test case to wait
for to use the promise-based wait functionality. If a promise is returned from
the `test` callback function, Jest will wait for that promise to resolve before
assuming all assertions are done. This also means that the rejected promises will
be passed to Jest, instead of being caughted by the global unhandledRejection handler.

Because the test will wait for the middleware promise, this means we can remove the
assertion expectation, since those will either run or the test will fail due to a
rejected promise.